+++ Code/Scripts/Test_DAK.cpp
@@ -127,7 +127,7 @@ DECLARE_SCRIPT(DAK_Vehicle_Regen_DAK, "" )
 		Commands->Send_Custom_Event ( obj, obj, 0, 0, 0 );
 	}
 
-	void Custom (GameObject* obj, int type, int param, GameObject* sender) override
+	void Custom (GameObject* obj, int type, uintptr_t param, GameObject* sender) override
 	{
 		if ( type == 0 ) // regenerate health.
 		{
@@ -155,7 +155,7 @@ DECLARE_SCRIPT(DAK_Electric_Death_DAK, "" )
 		Commands->Send_Custom_Event ( obj, obj, 1, 0, 1 ); // wait a second before applying next ammount of damage.
 	}
 
-	void Custom ( GameObject *obj, int type, int param, GameObject *sender ) override
+	void Custom ( GameObject *obj, int type, uintptr_t param, GameObject *sender ) override
 	{
 		if ( type == 0 ) // create next soldier, attach script, kill yourself with electric damage.
 		{
